delete book implemented

+++ Client/src/Components/Body/02-Bookcard.jsx
@@ -2,10 +2,23 @@ import React, { useEffect, useState } from "react";
 import Axios from "axios";
 import { updatebook } from "./bookSlice";
 import { useDispatch } from "react-redux";
+import DeletePopUp from "./04-DeletePopUp";
+import { createPortal } from "react-dom";
 export default function Bookcard({ data }) {
     let dispatch = useDispatch();
 
-    const deleteBook = (deleteId) => {
+    let [showPop, setShowPop] = useState(false);
+    let [deleteBook, setDeleteBook] = useState(false);
+    let [deleteId, setDeleteId] = useState("");
+
+    const deleteRecord = (options) => {};
+
+    const deleteBookfromDB = (id) => {
+        setShowPop(true);
+        setDeleteId(id);
+    };
+
+    if (deleteBook) {
         let options = {
             method: "delete",
             params: {
@@ -16,8 +29,9 @@ export default function Bookcard({ data }) {
 
         Axios.request(options).then((res) => {
             dispatch(updatebook());
+            setDeleteBook(false);
         });
-    };
+    }
 
     return (
         <div className="book-card">
@@ -31,9 +45,20 @@ export default function Bookcard({ data }) {
                 <p>Rent: {data.rent}</p>
                 <p>Buy: {data.buy}</p>
                 <button className="edit-btn modify">Edit</button>
-                <button className="delete-btn modify" onClick={() => deleteBook(data._id)}>
+                <button className="delete-btn modify" onClick={() => deleteBookfromDB(data._id)}>
                     Delete
                 </button>
+                {showPop
+                    ? createPortal(
+                          <DeletePopUp
+                              showPop={showPop}
+                              setShowPop={setShowPop}
+                              deleteBook={deleteBook}
+                              setDeleteBook={setDeleteBook}
+                          />,
+                          document.getElementById("root")
+                      )
+                    : null}
             </div>
         </div>
     );
